+++ TouchGFX/gui/include/gui/containers/SettingsContainter.hpp
@@ -11,7 +11,7 @@ public:
     void initialize();
 
     void UpdateSettingsLabel(char* name, uint8_t len);
-    void UpdateSettingsValue(int8_t value);
+    void UpdateSettingsValue(int16_t value);
 protected:
 };
